@@ -1,9 +1,8 @@
 package main
 
 import (
-	
 	"fmt"
-	"io/ioutil"
+	"os"
 	"strings"
 
 	"github.com/astaxie/beego/migration"
@@ -25,7 +24,7 @@ func init() {
 // Run the migrations
 func (m *CrearEsquemaActaRecibido_20190723_112831) Up() {
 	// use m.SQL("CREATE TABLE ...") to make schema update
-	file, err := ioutil.ReadFile("../scripts/20190723_112831_crear_esquema_acta_recibido_up.sql")
+	file, err := os.ReadFile("../scripts/20190723_112831_crear_esquema_acta_recibido_up.sql")
 
 	if err != nil {
 		// handle error
@@ -44,7 +43,7 @@ func (m *CrearEsquemaActaRecibido_20190723_112831) Up() {
 // Reverse the migrations
 func (m *CrearEsquemaActaRecibido_20190723_112831) Down() {
 	// use m.SQL("DROP TABLE ...") to reverse schema update
-	file, err := ioutil.ReadFile("../scripts/20190723_112831_crear_esquema_acta_recibido_down.sql")
+	file, err := os.ReadFile("../scripts/20190723_112831_crear_esquema_acta_recibido_down.sql")
 
 	if err != nil {
 		// handle error
