@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
+	"os"
 	"strings"
 
 	"github.com/astaxie/beego/migration"
@@ -23,7 +23,7 @@ func init() {
 
 // Run the migrations
 func (m *RegistroInmuebles_20221222_011056) Up() {
-	file, err := ioutil.ReadFile("../scripts/20221222_011056_registro_inmuebles_up.sql")
+	file, err := os.ReadFile("../scripts/20221222_011056_registro_inmuebles_up.sql")
 
 	if err != nil {
 		// handle error
@@ -40,7 +40,7 @@ func (m *RegistroInmuebles_20221222_011056) Up() {
 
 // Reverse the migrations
 func (m *RegistroInmuebles_20221222_011056) Down() {
-	file, err := ioutil.ReadFile("../scripts/20221222_011056_registro_inmuebles_down.sql")
+	file, err := os.ReadFile("../scripts/20221222_011056_registro_inmuebles_down.sql")
 
 	if err != nil {
 		// handle error
